fix: playback rate truncation, null-check dlopen/CFBundleCreate, update CI
Fix playback rate bug where double values (e.g. 1.5x) were truncated
to int, producing incorrect elapsed time calculations.

Add return value checks for dlopen() and CFBundleCreate() with early
exit on failure. Replace mixed @autoreleasepool / manual
NSAutoreleasePool with a single @autoreleasepool block.

Update CI workflow: simplify get-raw output test to just check exit
code (output format varies by macOS version), update runner matrix
to macos-13/14/15 (11/12 are deprecated).

+++ src/mediaremote-mini/adapter/now_playing.m
@@ -35,14 +35,14 @@ NSNumber *getElapsedTimeNow(NSDictionary *information) {
     NSTimeInterval currentEpoch = [[NSDate date] timeIntervalSince1970];
     NSTimeInterval timeDiff = currentEpoch - timestampEpoch;
 
-    int playbackRate = 0;
+    double playbackRate = 0.0;
     id playbackRateVal = information[kMRMediaRemoteNowPlayingInfoPlaybackRate];
     if ([playbackRateVal isKindOfClass:[NSNumber class]]) {
         playbackRate = [(NSNumber *)playbackRateVal doubleValue];
     }
 
     double realElapsed = [(NSNumber *)elapsed doubleValue];
-    if (playbackRate >= 0) {
+    if (playbackRate >= 0.0) {
         realElapsed += timeDiff * playbackRate;
     }
